requirements.txt
python-dotenv
openai
pyupbit
ta
selenium
webdriver-manager
Pillow
youtube-transcript-api